climate-tracker-app/backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Flask Backend (Alternative/Additional)
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External APIs & Data Processing
aiohttp==3.9.1
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# Image Processing & Storage
pillow==10.1.0
cloudinary==1.36.0
boto3==1.34.0

# ML/AI
scikit-learn==1.3.2
joblib==1.3.2
opencv-python==4.8.1.78

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
structlog==23.2.0

# Production
gunicorn==21.2.0